new branch HBase

+++ src/main/java/task2/ChainMapper2.java
@@ -1,6 +1,7 @@
 package task2;
 
 import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.mapreduce.Mapper;
 
 import java.io.IOException;
@@ -9,16 +10,16 @@ import java.util.*;
 /**
  * Created by fotis on 01/02/16.
  */
-public class ChainMapper2 extends Mapper<IntWritable, IntWritable, IntWritable, IntWritable> {
+public class ChainMapper2 extends Mapper<LongWritable, LongWritable, LongWritable, LongWritable> {
     private PriorityQueue<CustomPairMods> pq = new PriorityQueue<CustomPairMods>();
     private int k;
 
     public void setup (Context context){
         k = Integer.parseInt(context.getConfiguration().get("k"));
     }
-    public void map(IntWritable key, IntWritable value, Context context) throws IOException,InterruptedException {
-        int articleId = key.get();
-        int modifications =value.get();
+    public void map(LongWritable key, LongWritable value, Context context) throws IOException,InterruptedException {
+        long articleId = key.get();
+        long modifications =value.get();
         CustomPairMods customPairMods = new CustomPairMods(articleId, modifications);
         pq.add(customPairMods);
         if (pq.size() > k){
@@ -32,7 +33,7 @@ public class ChainMapper2 extends Mapper<IntWritable, IntWritable, IntWritable, 
         CustomPairMods customPairMods;
         while(pq.size() > 0){
             customPairMods = pq.poll();
-            context.write(new IntWritable(customPairMods.getArticleId()), new IntWritable(customPairMods.getModifications()));
+            context.write(new LongWritable(customPairMods.getArticleId()), new LongWritable(customPairMods.getModifications()));
         }
     }
 }
