+++ src/main/java/task2/ReducerTask2.java
@@ -1,6 +1,7 @@
 package task2;
 
 import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
@@ -10,15 +11,15 @@ import java.io.IOException;
 /**
  * Created by fotis on 01/02/16.
  */
-public class ReducerTask2 extends Reducer<IntWritable, IntWritable, IntWritable, IntWritable> {
+public class ReducerTask2 extends Reducer<LongWritable, LongWritable, LongWritable, LongWritable> {
 
-    public void reduce(IntWritable key, Iterable<IntWritable> values, Context context) throws IOException,InterruptedException {
+    public void reduce(LongWritable key, Iterable<LongWritable> values, Context context) throws IOException,InterruptedException {
 
         int sum = 0;
-        for (IntWritable val : values){
+        for (LongWritable val : values){
             sum+=val.get();
         }
-        context.write(key, new IntWritable(sum));
+        context.write(key, new LongWritable(sum));
 
     }
